博客编辑修改

+++ DimpleBlog-master/src/main/java/com/dimple/project/blog/blog/mapper/BlogMapper.java
@@ -206,5 +206,5 @@ public interface BlogMapper  extends BaseMapper<Blog> {
     List<Blog> selectBlogsByCreator(Blog blog);
 
     
-	List<Blog> selectBlogListByFuncId(@Param("funcId")Integer funcId);
+	List<Blog> selectBlogListByFuncId(@Param("funcId")Long funcId);
 }
